actions/review-checklists/requirements.txt
PyGithub>=2.1.1
PyYAML>=6.0.1

